edited stlye of nav buttons when logged in and moved profile page to pages folder
